requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
docopt==0.6.2
gitdb2==2.0.3
GitPython==2.1.7
idna==2.6
influxdb==4.1.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
smmap2==2.0.3
urllib3==1.22
